refactor: Extract expandRoutes
